src/webapp/requirements.txt
gradio==3.50.1
requests
pillow